+++ src/activeNetworkInterfaceModel.cpp
@@ -86,6 +86,7 @@ ActiveNetworkInterfaceModel::ActiveNetworkInterfaceModel(QObject* parent)
 
 ActiveNetworkInterfaceModel::~ActiveNetworkInterfaceModel()
 {
+	// Remove settings observers
 	auto& settings = settings::SettingsManager::getInstance();
 	settings.unregisterSettingObserver(settings::InterfaceTypeWiFi.name, d_ptr.get());
 	settings.unregisterSettingObserver(settings::InterfaceTypeEthernet.name, d_ptr.get());